requirements.txt
docx2text
ipykernel
python-dotenv
llama-index

qdrant_client